requirements.txt
scanpy>=1.9.3
scvi-tools>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.5.0
panel>=1.0.0
bokeh>=2.4.0
transformers>=4.52.3
openai>=1.82.0
dotenv